[main] Revamp all headings
I sacrificed the ability to put literally anything as a heading in favor
of being able to implement TOC generation in the future.

On the positive side, headings are now even easier to write, and they
span only one line of code, which makes the source code a lot more
human-readable.