[main] Add 2nd form of heading macros
The former use of headings is still the same (except the Hf macro got
renamed to NHf). The new form of headings uses the SH and NH macros as
starting markers, and the newly introduced HE macro as the ending
marker. This makes it possible to include arbitrarily complex elements
as parts of the heading (though the main motivation was to be able to
use URLs in headings). See updated README for more information.