Fix a blatant lie inside README
The reason rnd macros don't automatically source other files is because
there is no guarantee these other macro files have not been modified by
some user, and most people have automated build scripts for compiling
Neatroff documents, so it's not an issue to have to add a few more -m
options in the command.
Also, the ".so" request doesn't always seem to work as expected when I
use it and I couldn't be bothered to investigate this, could be an issue
of using symlinks.